Initial commit
streaming.py is the back end and converts the data from the index.html
page into the direction commands for the robot.

It also captures the live video feed and passes the frames as jpeg files
to the html page to display as a live stream video

index.html has a live stream feed from the camera device on the front if
the robot is going forward once back is pressed the robot uses the back
camera until forward is pressed again.